context: refactor context uses for consistency
many builtin functions rely on the PitContext object to feed them
pathnames to various directories on the repository (e.g, the .git/ repo,
the working tree directory, etc.)

this PitContext object is passed as first argument to those functions,
but it's less than ideal since we need, in the future, to pass the list
of args from the command line, and this ctx object gets in the way.

remove the ctx argument in call sites to those functions, and instead
make them get the context themselves by calling context.get_context()

get_context() looks for the .git repository in the current dir, and if
it doesn't find any, looks recursively in the parent directories until
the root directory.

Signed-off-by: Sébastien Guimmara <sebastien.guimmara@gmail.com>