hash_file: return SHA-1 hexdigest instead of digest
hash_file returns the 20-byte SHA-1. Make it return the 40 bytes
hex SHA-1 instead.

This adds a bit of overhead because we need to convert it back to
20-bytes SHA-1 whenever we write it on disk, but this improves
readability in test output and it increases consistency with other
hashing functions.

Since this is a pedagogical implementation, the main goal is readability
over efficiency.

Signed-off-by: Sébastien Guimmara <sebastien.guimmara@gmail.com>